+++ src/containerapp/azext_containerapp/_clients.py
@@ -993,6 +993,7 @@ class JavaComponentPreviewClient():
 
         return java_component_list
 
+
 class SessionPoolPreviewClient():
     api_version = PREVIEW_API_VERSION
 
@@ -1009,40 +1010,6 @@ class SessionPoolPreviewClient():
             cls.api_version)
 
         r = send_raw_request(cmd.cli_ctx, "PUT", request_url, body=json.dumps(session_pool_envelope))
-
-        # try to add user as session pool creator role to the session pool
-        try:
-            # get princpalId of the user
-            principal_id_url = "https://graph.microsoft.com/v1.0/me"
-            principal_id = send_raw_request(cmd.cli_ctx,"GET",principal_id_url).json()['id']
-            scope = "subscriptions/{}/resourceGroups/{}/providers/Microsoft.App/sessionPools/{}".format(sub_id, resource_group_name, name)
-            role_assignment_fmt = "{}/{}/providers/Microsoft.Authorization/roleAssignments/{}?api-version=2022-04-01"
-            role_assignment_url = role_assignment_fmt.format(
-                management_hostname.strip('/'),
-                scope,
-                uuid.uuid4()
-            )
-            role_definition_id = "/{}/providers/Microsoft.Authorization/roleDefinitions/{}".format(
-                scope,
-                SESSION_CREATOR_ROLE_ID)
-            assign_role_r = send_raw_request(cmd.cli_ctx,"PUT", role_assignment_url, body=json.dumps({
-                "properties": {
-                    "roleDefinitionId": role_definition_id,
-                    "principalId": principal_id
-                }
-            }))
-        except Exception as e:
-            try:
-                if isinstance(e, HTTPError):
-                    error_code = json.loads(e.response.text)["error"]["code"]
-                    if error_code == "RoleAssignmentExists":
-                        pass
-                else:
-                    logger.warning("Could not add user as session pool creator role to the session pool, please follow the docs https://learn.microsoft.com/en-us/azure/container-apps/sessions-code-interpreter?tabs=azure-cli#authentication to add the needed roll for authentication")
-                    logger.warning(e)
-            except:
-                logger.warning("Could not add user as session pool creator role to the session pool, please follow the docs https://learn.microsoft.com/en-us/azure/container-apps/sessions-code-interpreter?tabs=azure-cli#authentication to add the needed roll for authentication")
-                logger.warning(e)
 
         if no_wait:
             return r.json()
@@ -1165,19 +1132,24 @@ class SessionPoolPreviewClient():
 
         return sessionpools
 
-class SessionCodeInterpreterPreviewClient():
+
+class DotNetComponentPreviewClient():
     api_version = PREVIEW_API_VERSION
 
     @classmethod
-    def execute(cls, cmd, identifier, code_interpreter_envelope, session_pool_endpoint, no_wait=False):
-        url_fmt = "{}/code/execute?identifier={}&api-version={}"
+    def create(cls, cmd, resource_group_name, environment_name, name, dotnet_component_envelope, no_wait=False):
+        management_hostname = cmd.cli_ctx.cloud.endpoints.resource_manager
+        sub_id = get_subscription_id(cmd.cli_ctx)
+        url_fmt = "{}/subscriptions/{}/resourceGroups/{}/providers/Microsoft.App/managedEnvironments/{}/dotNetComponents/{}?api-version={}"
         request_url = url_fmt.format(
-            session_pool_endpoint,
-            identifier,
-            PREVIEW_API_VERSION)
-        logger.warning(request_url)
-        logger.warning(code_interpreter_envelope)
-        r = send_raw_request(cmd.cli_ctx, "POST", request_url, body=json.dumps(code_interpreter_envelope), resource=SESSION_RESOURCE)
+            management_hostname.strip('/'),
+            sub_id,
+            resource_group_name,
+            environment_name,
+            name,
+            cls.api_version)
+
+        r = send_raw_request(cmd.cli_ctx, "PUT", request_url, body=json.dumps(dotnet_component_envelope))
 
         if no_wait:
             return r.json()
@@ -1189,103 +1161,89 @@ class SessionCodeInterpreterPreviewClient():
         return r.json()
 
     @classmethod
-    def upload(cls,cmd,identifier,filepath,session_pool_endpoint, no_wait=False):
-        url_fmt = "{}/files/upload?identifier={}&api-version={}"
+    def update(cls, cmd, resource_group_name, environment_name, name, dotnet_component_envelope, no_wait=False):
+        management_hostname = cmd.cli_ctx.cloud.endpoints.resource_manager
+        sub_id = get_subscription_id(cmd.cli_ctx)
+        url_fmt = "{}/subscriptions/{}/resourceGroups/{}/providers/Microsoft.App/managedEnvironments/{}/dotNetComponents/{}?api-version={}"
         request_url = url_fmt.format(
-            session_pool_endpoint,
-            identifier,
-            PREVIEW_API_VERSION)
+            management_hostname.strip('/'),
+            sub_id,
+            resource_group_name,
+            environment_name,
+            name,
+            cls.api_version)
 
-        from azure.cli.core._profile import Profile
-        profile = Profile(cli_ctx=cmd.cli_ctx)
-        token_info, _, _ = profile.get_raw_token(resource=SESSION_RESOURCE)
-        _, token, _ = token_info
-        headers = {'Authorization': 'Bearer ' + token}
-        
-        try:
-            data_file = open(filepath, "rb")
-            file_name = os.path.basename(filepath)
-            files = [("file", (file_name, data_file))]
-
-            r = requests.post(
-                request_url,
-                files=files,
-                headers=headers)
-            
-            data_file.close()
-        except Exception as e:
-            logger.error("error occurred while uploading file")
-            return str(e)
+        r = send_raw_request(cmd.cli_ctx, "PATCH", request_url, body=json.dumps(dotnet_component_envelope))
 
         if no_wait:
-            return r.json()
-        elif r.status_code in [200, 201, 202, 204]:
-            logger.warning("upload success")
-            if r.status_code == 202:
-                operation_url = r.headers.get(HEADER_AZURE_ASYNC_OPERATION)
-                poll_status(cmd, operation_url)
-                r = send_raw_request(cmd.cli_ctx, "GET", request_url, resource=SESSION_RESOURCE)
+            return
+        elif r.status_code == 202:
+            operation_url = r.headers.get(HEADER_LOCATION)
+            response = poll_results(cmd, operation_url)
+            if response is None:
+                raise ResourceNotFoundError("Could not find the DotNet component")
+            else:
+                return response
 
         return r.json()
 
     @classmethod
-    def show_file_content(cls, cmd, identifier, filename, session_pool_endpoint, no_wait=False):
-        url_fmt = "{}/files/content/{}?identifier={}&api-version={}"
+    def delete(cls, cmd, resource_group_name, environment_name, name, no_wait=False):
+        management_hostname = cmd.cli_ctx.cloud.endpoints.resource_manager
+        sub_id = get_subscription_id(cmd.cli_ctx)
+        url_fmt = "{}/subscriptions/{}/resourceGroups/{}/providers/Microsoft.App/managedEnvironments/{}/dotNetComponents/{}?api-version={}"
         request_url = url_fmt.format(
-            session_pool_endpoint,
-            filename,
-            identifier,
-            PREVIEW_API_VERSION)
+            management_hostname.strip('/'),
+            sub_id,
+            resource_group_name,
+            environment_name,
+            name,
+            cls.api_version)
 
-        r = send_raw_request(cmd.cli_ctx, "GET", request_url, resource=SESSION_RESOURCE)
-        # print out the file content as bytes decoded as string
-        logger.warning(r.content.decode())
-
-        return json.dumps(r.content.decode())
-    
-    @classmethod
-    def show_file_metadata(cls, cmd, identifier, filename, session_pool_endpoint, no_wait=False):
-        url_fmt = "{}/files/{}?identifier={}&api-version={}"
-        request_url = url_fmt.format(
-            session_pool_endpoint,
-            filename,
-            identifier,
-            PREVIEW_API_VERSION)
-        logger.warning(request_url)
-        r = send_raw_request(cmd.cli_ctx, "GET", request_url, resource=SESSION_RESOURCE)
-
-        return r.json()
-
-    @classmethod
-    def delete_file(cls, cmd, identifier, filename, session_pool_endpoint, no_wait=False):
-        url_fmt = "{}/files/{}?identifier={}&api-version={}"
-        request_url = url_fmt.format(
-            session_pool_endpoint,
-            filename,
-            identifier,
-            PREVIEW_API_VERSION)
-
-        r = send_raw_request(cmd.cli_ctx, "DELETE", request_url, resource=SESSION_RESOURCE)
+        r = send_raw_request(cmd.cli_ctx, "DELETE", request_url)
 
         if no_wait:
             return  # API doesn't return JSON (it returns no content)
         elif r.status_code in [200, 201, 202, 204]:
-            logger.warning('file successfully deleted')
             if r.status_code == 202:
                 operation_url = r.headers.get(HEADER_LOCATION)
                 poll_results(cmd, operation_url)
 
     @classmethod
-    def list_files(cls, cmd, identifier, path, session_pool_endpoint, no_wait=False):
-        if path is None:
-            path = ""
-            
-        url_fmt = "{}/files?identifier={}&path={}&api-version={}"
+    def show(cls, cmd, resource_group_name, environment_name, name):
+        management_hostname = cmd.cli_ctx.cloud.endpoints.resource_manager
+        sub_id = get_subscription_id(cmd.cli_ctx)
+        url_fmt = "{}/subscriptions/{}/resourceGroups/{}/providers/Microsoft.App/managedEnvironments/{}/dotNetComponents/{}?api-version={}"
         request_url = url_fmt.format(
-            session_pool_endpoint,
-            identifier,
-            path,
-            PREVIEW_API_VERSION)
-        
-        r = send_raw_request(cmd.cli_ctx, "GET", request_url, resource=SESSION_RESOURCE)
+            management_hostname.strip('/'),
+            sub_id,
+            resource_group_name,
+            environment_name,
+            name,
+            cls.api_version)
+
+        r = send_raw_request(cmd.cli_ctx, "GET", request_url)
+
         return r.json()
+
+    @classmethod
+    def list(cls, cmd, resource_group_name, environment_name):
+        dotNet_component_list = []
+
+        management_hostname = cmd.cli_ctx.cloud.endpoints.resource_manager
+        sub_id = get_subscription_id(cmd.cli_ctx)
+        url_fmt = "{}/subscriptions/{}/resourceGroups/{}/providers/Microsoft.App/managedEnvironments/{}/dotNetComponents?api-version={}"
+        request_url = url_fmt.format(
+            management_hostname.strip('/'),
+            sub_id,
+            resource_group_name,
+            environment_name,
+            cls.api_version)
+
+        r = send_raw_request(cmd.cli_ctx, "GET", request_url)
+        r = r.json()
+
+        for component in r["value"]:
+            dotNet_component_list.append(component)
+
+        return dotNet_component_list