+++ src/containerapp/azext_containerapp/_models.py
@@ -615,3 +615,9 @@ CustomDomainConfiguration = {
 RuntimeJava = {
     "enableMetrics": False
 }
+
+DotNetComponent = {
+    "properties": {
+        "componentType": "AspireDashboard"
+    }
+}