+++ src/serviceconnector-passwordless/azext_serviceconnector_passwordless/_params.py
@@ -15,7 +15,9 @@ from azure.cli.command_modules.serviceconnector._params import (
     add_connection_string_argument,
     add_secret_store_argument,
     add_local_connection_block,
-    add_customized_keys_argument
+    add_customized_keys_argument,
+    add_configuration_store_argument,
+    add_opt_out_argument
 )
 from azure.cli.command_modules.serviceconnector._validators import (
     get_default_object_id_of_current_user
@@ -61,7 +63,7 @@ def load_arguments(self, _):
             add_target_resource_block(c, target)
             add_auth_block(c, source, target)
             add_new_addon_argument(c, source, target)
-            add_secret_store_argument(c)
+            add_secret_store_argument(c, source)
             add_vnet_block(c, target)
             add_local_connection_block(c)
             add_customized_keys_argument(c)
@@ -75,9 +77,11 @@ def load_arguments(self, _):
                 add_source_resource_block(c, source, enable_id=False)
                 add_target_resource_block(c, target)
                 add_auth_block(c, source, target)
+                add_configuration_store_argument(c)
                 add_new_addon_argument(c, source, target)
-                add_secret_store_argument(c)
+                add_secret_store_argument(c, source)
                 add_vnet_block(c, target)
                 add_connection_string_argument(c, source, target)
                 add_customized_keys_argument(c)
+                add_opt_out_argument(c)
                 c.argument('yes', arg_type=yes_arg_type)